profile done

+++ app/Http/Controllers/API/UserController.php
@@ -12,6 +12,16 @@ use Illuminate\Support\Facades\Auth;
 class UserController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth:api');
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -64,6 +74,42 @@ class UserController extends Controller
         //
     }
 
+    public function profile()
+    {
+        return auth('api')->user();
+    }
+
+    public function updateProfile(Request $request)
+    {
+        $user = auth('api')->user();
+        $currentPhoto = $user->photo;
+
+        $this->validate($request,[
+            'name' => 'required|string|max:191',
+            'email' => 'required|string|email|max:191|unique:users,email,'.$user->id,
+            'password' => 'sometimes|required|min:8'
+        ]);
+
+        if($request->photo != $currentPhoto) {
+            $name = time().'.' . explode('/', explode(':', substr($request->photo, 0, strpos($request->photo, ';')))[1])[1];
+            
+            \Image::make($request->photo)->save(public_path('img/profile/').$name);
+            $request->merge(['photo' => $name]);
+
+            $userPhoto = public_path('img/profile/').$currentPhoto;
+            if(file_exists($userPhoto)){
+                @unlink($userPhoto);
+            }
+        }
+
+        if(!empty($request->password)){
+            $request->merge(['password' => Hash::make($request['password'])]);
+        }
+
+        $user->update($request->all());
+        return ['message' => 'Profile updated successfully'];
+    }
+
     /**
      * Update the specified resource in storage.
      *
@@ -94,21 +140,22 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
-        $currentUser = Auth::user();
+        $currentUser = auth('api')->user();
         $user = User::findOrFail($id);
 
         // return ['message' => $currentUser]; /* this returns null! */
 
-        // if($currentUser->id == $user->id) {
-        //     return ['message' => `can't delete the current user`];
+        if($currentUser->id == $user->id) {
 
-        // } else {
-        //     // $user->delete();
+            return response()->json(['message' => 'can not delete the current user'], 403);
+
+        } else {
+            $user->delete();
         
-        //     return ['message' => 'User deleted successfuly'];
-        // }
+            return ['message' => 'User deleted successfuly'];
+        }
 
-        $user->delete();
-        return ['message' => 'User deleted successfully'];
+        // $user->delete();
+        // return ['message' => 'User deleted successfully'];
     }
 }
